requirements.txt
fastapi==0.115.0

uvicorn[standard]==0.30.6

pydantic==2.8.2

typing-extensions>=4.8,<5.0

psycopg==3.2.1

psycopg-pool==3.2.2

python-dotenv==1.0.1

python-telegram-bot==20.7
